+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.1.0
+
+### Configuration Scope
+
+- Added the following commands:
+  - `spanningTree.enableRapidPVST`
+  - `spanningTree.portfast.disable`
+  - `spanningTree.portfast.bpduGuard.disable`
+  - `spanningTree.portfast.bpduGuard.enable`
+  - `spanningTree.portfast.enable`
+  - `spanningTree.portfast.enableForTrunk`
+  - `spanningTree.setPriorityValue`
+  - `spanningTree.setPriorityValueByOffset`
+
 ## 1.0.1
 
 ### Bugfixes